requirements.txt
streamlit
supabase
python-dotenv
requests
pandas
plotly
